refactor(data-model): reuse shared helpers, trim narrative comments
- Drop duplicate getCompositeSubFieldLabel + COMPOSITE_SUB_FIELD_LABELS;
  use the existing helper under object-record/object-filter-dropdown
  (backed by SETTINGS_COMPOSITE_FIELD_TYPE_CONFIGS).
- Move MAX_CUSTOM_INDEXES_PER_OBJECT to twenty-shared/constants so the
  cap is defined once for FE + BE.
- Use isNonEmptyString (@sniptt/guards) in place of inline
  isDefined(x) && x !== '' guards.
- Build a fieldsById Map in the index table builder; narrow the useMemo
  dep to indexMetadatas + fields.
- Stabilize indexableFields with useMemo in SettingsObjectNewIndex.
- Drop the runtime isUnique guard + corresponding test + enum value: the
  DTO doesn't expose isUnique, GraphQL rejects unknown fields, and the
  test only exercised an `as any` bypass.
- Trim narrative / WHAT / PR-history comments across the changed files.
