--- packages/twenty-front/src/modules/settings/data-model/indexes/utils/getCompositeSubFieldLabel.ts
@@ -1,8 +1,0 @@
-import { COMPOSITE_SUB_FIELD_LABELS } from '@/settings/data-model/indexes/constants/CompositeSubFieldLabels';
-import { type FieldMetadataType } from '~/generated-metadata/graphql';
-
-export const getCompositeSubFieldLabel = (
-  parentType: FieldMetadataType,
-  subFieldName: string,
-): string =>
-  COMPOSITE_SUB_FIELD_LABELS[parentType]?.[subFieldName] ?? subFieldName;
