--- packages/twenty-front/src/pages/settings/data-model/constants/MaxCustomIndexesPerObject.ts
@@ -1,4 +1,0 @@
-// Mirrors MAX_CUSTOM_INDEXES_PER_OBJECT on the server. Kept in sync manually —
-// the server is the source of truth and will reject anything above the cap;
-// this constant just lets the UI disable the "Add Index" button preemptively.
-export const MAX_CUSTOM_INDEXES_PER_OBJECT = 10;
