+++ packages/twenty-server/src/engine/metadata-modules/index-metadata/index-field-metadata.entity.ts
@@ -57,10 +57,9 @@ export class IndexFieldMetadataEntity implements Required<IndexFieldMetadataEnti
   @Column({ nullable: false })
   order: number;
 
-  // Composite-type fields (Address, Currency, ...) don't map to a single SQL
-  // column. When the user picks "Address > City", we store the parent
-  // fieldMetadataId here AND the composite sub-property name in subFieldName
-  // (e.g. 'addressCity'). For scalar/relation fields this is null.
+  // Null for scalar/relation fields. Set to the composite sub-property name
+  // (e.g. 'addressCity') when the index targets a single column of a
+  // composite-type parent.
   @Column({ type: 'text', nullable: true })
   subFieldName: string | null;
