+++ packages/twenty-server/src/engine/metadata-modules/index-metadata/index-field-metadata.exception.ts
@@ -27,5 +27,4 @@ export enum IndexMetadataExceptionCode {
   INDEX_OBJECT_NOT_FOUND = 'INDEX_OBJECT_NOT_FOUND',
   INDEX_FIELD_NOT_FOUND_ON_OBJECT = 'INDEX_FIELD_NOT_FOUND_ON_OBJECT',
   INDEX_NOT_FOUND = 'INDEX_NOT_FOUND',
-  UNIQUE_INDEX_NOT_ALLOWED = 'UNIQUE_INDEX_NOT_ALLOWED',
 }
