+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-index-metadata.type.ts
@@ -4,8 +4,6 @@ import { type IndexMetadataEntity } from 'src/engine/metadata-modules/index-meta
 import { type UniversalFlatEntityFrom } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-entity-from.type';
 
 // Note: IndexFieldMetadataEntity is just in between a SyncableEntity and a jsonb we should decide one of both https://github.com/twentyhq/core-team-issues/issues/2227
-// subFieldName carries through automatically since it's a real column on the
-// entity (not stripped by the Omit list).
 export type UniversalFlatIndexFieldMetadata = Omit<
   IndexFieldMetadataEntity,
   | 'indexMetadataId'
