+++ packages/twenty-shared/src/constants/MaxCustomIndexesPerObject.ts
@@ -1,0 +1,1 @@
+export const MAX_CUSTOM_INDEXES_PER_OBJECT = 10;
